+++ src/components/layout/Sidebar/index.ts
@@ -1,0 +1,3 @@
+export { Sidebar } from "./Sidebar.container";
+export { SidebarUI } from "./Sidebar.ui";
+export type { SidebarUIProps } from "./Sidebar.ui";
